requirements.txt
requests>=2.25.1
opencv-python>=4.5.1
numpy>=1.20.1
pytesseract>=0.3.8
Pillow>=8.2.0
selenium>=4.1.0
webdriver-manager>=3.5.2
Flask>=2.0.1
Flask-CORS>=3.0.10